[todos] add redux, make 'Add Task' button add 'whats up' todo
