+++ package.json
@@ -8,6 +8,7 @@
     "@fortawesome/react-fontawesome": "^0.1.3",
     "bulma": "^0.7.2",
     "classnames": "^2.2.6",
+    "flux": "^3.1.3",
     "i": "^0.3.6",
     "moment": "^2.22.2",
     "node-sass": "^4.10.0",
@@ -15,7 +16,9 @@
     "react": "^16.6.0",
     "react-dom": "^16.6.0",
     "react-fontawesome": "^1.6.1",
-    "react-scripts": "2.1.1"
+    "react-redux": "^5.1.1",
+    "react-scripts": "2.1.1",
+    "redux": "^4.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
