+++ src/App.js
@@ -1,28 +1,32 @@
 import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import Inbox from './Inbox';
 import Today from './Today';
 import Menu from './Menu';
+import { TodoType } from './TodoItem';
+import { connect } from 'react-redux';
 import './App.sass';
 
-const todos = [
-  {name: 'do laundry'},
-  {name: 'write some code'},
-  {name: 'clean my room'},
-];
-
 class App extends Component {
+  static propTypes = {
+    todos: PropTypes.arrayOf(TodoType).required,
+  };
+
   state = {
     selected: 'Inbox',
   };
 
   render() {
+    const {todos} = this.props;
     const {selected} = this.state;
+
     var content;
     if (selected === 'Today') {
       content = <Today />;
     } else {
       content = <Inbox todos={todos} />;
     }
+
     return (
       <div className="App">
         <div className="todo-navbar navbar" >
@@ -45,4 +49,10 @@ class App extends Component {
   }
 }
 
-export default App;
+const mapStateToProps = function(state) {
+  return {
+    todos: state.todos,
+  };
+}
+
+export default connect(mapStateToProps)(App);
