+++ src/actions.js
@@ -1,0 +1,4 @@
+export const addTodo = (name) => ({
+  name,
+  type: 'ADD_TODO'
+});
